Move logs to contracts
