+++ .vscode/settings.json
@@ -12,6 +12,7 @@
   },
   "typescript.tsdk": "node_modules/typescript/lib",
   "typescript.enablePromptUseWorkspaceTsdk": true,
+  "typescript.tsserver.experimental.enableProjectDiagnostics": true,
   "sqltools.connections": [
     {
       "previewLimit": 50,
