+++ packages/contracts/prettier.config.cjs
@@ -1,0 +1,10 @@
+/**
+ * @see https://prettier.io/docs/configuration
+ * @type {import("prettier").Config}
+ */
+const config = {
+  singleQuote: true,
+  semi: false,
+}
+
+module.exports = config
