--- packages/contracts/prettier.config.ts
@@ -1,8 +1,0 @@
-import { type Config } from "prettier";
-
-const config: Config = {
-  singleQuote: true,
-  semi: false,
-};
-
-export default config;
