+++ packages/contracts/src/settings/logs/index.ts
@@ -1,2 +1,4 @@
+export * from './logEvent.dto'
+export * from './logFile.dto'
 export * from './logSetting'
 export * from './logSetting.dto'
