+++ packages/contracts/src/settings/logs/logEvent.dto.ts
@@ -1,0 +1,5 @@
+export type LogEvent = {
+  message: string
+  date: Date
+  level: string
+}
