+++ packages/contracts/src/settings/logs/logFile.dto.ts
@@ -1,0 +1,4 @@
+export type LogFile = {
+  name: string
+  size: number
+}
