--- server/src/modules/logging/dtos/logFile.dto.ts
@@ -1,4 +1,0 @@
-export type LogFile = {
-  name: string;
-  size: number;
-};
