--- server/src/modules/logging/logEvent.ts
@@ -1,5 +1,0 @@
-export type LogEvent = {
-  message: string;
-  date: Date;
-  level: string;
-};
